requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.5.0
dj-email-url==1.0.5
django-allauth==0.50.0
django-cache-url==3.4.0
django-ckeditor-5==0.0.16
django-crispy-forms==1.14.0
django==4.0.4
environs[django]==9.5.0
idna==3.3
marshmallow==3.15.0
oauthlib==3.2.0
packaging==21.3; python_version >= '3.6'
pillow==9.1.0
psycopg2-binary==2.9.3
pycparser==2.21
pyjwt[crypto]==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
python3-openid==3.2.0
requests-oauthlib==1.3.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
webencodings==0.5.1
